+++ Changes.md
@@ -4,11 +4,13 @@
 - NEW: `Parse::Model.autosave_on_create` has been removed in favor of `first_or_create!`.
 - NEW: Hooks and Functions now have a `log` method, similar to `puts`, but allows easier tracing of
 single requests in a multi-request threaded environment. (See Parse::Webhooks::Payload)
+- NEW: `:id` constraints also safely supports pointers by skipping class matching.
 - FIXED: first_or_create will now apply dirty tracking to newly created fields.
 - FIXED: Properties of :array type will always return a Parse::CollectionProxy if their internal value is nil. The object will not be marked dirty until something is added to the array.
 - FIXED: Encoding a Parse::Object into JSON will remove any values that are `nil`
 which were not explicitly changed to that value.
 - [PR#39](https://github.com/modernistik/parse-stack/pull/39): Allow Moneta::Expires as cache object to allow for non-native expiring caches by [GrahamW](https://github.com/GrahamW)
+
 
 ### 1.7.1
 - NEW: `:timezone` datatype that maps to `Parse::TimeZone` (which mimics `ActiveSupport::TimeZone`)